Make conversation link available for copy and paste in the Conversation settings.
