+++ app/src/main/java/com/nextcloud/talk/utils/ShareUtils.kt
@@ -6,7 +6,10 @@
  */
 package com.nextcloud.talk.utils
 
+import android.app.Activity
 import android.content.Context
+import android.content.Intent
+import android.net.Uri
 import com.nextcloud.talk.R
 import com.nextcloud.talk.data.user.model.User
 import com.nextcloud.talk.models.domain.ConversationModel
@@ -19,5 +22,35 @@ object ShareUtils {
             user.baseUrl,
             conversation?.token
         )
+    }
+
+    fun shareConversationLink(context: Activity, baseUrl: String?, roomToken: String?, conversationName: String?) {
+        if (baseUrl.isNullOrBlank() || roomToken.isNullOrBlank() || conversationName.isNullOrBlank()) {
+            return
+        }
+
+        val uriToShareConversation = Uri.parse(baseUrl)
+            .buildUpon()
+            .appendPath("index.php")
+            .appendPath("call")
+            .appendPath(roomToken)
+            .build()
+
+        val shareConversationLink = String.format(
+            context.getString(
+                R.string.share_link_to_conversation,
+                conversationName,
+                uriToShareConversation.toString()
+            )
+        )
+
+        val sendIntent: Intent = Intent().apply {
+            action = Intent.ACTION_SEND
+            putExtra(Intent.EXTRA_TEXT, shareConversationLink)
+            type = "text/plain"
+        }
+
+        val shareIntent = Intent.createChooser(sendIntent, context.getString(R.string.nc_share_link))
+        context.startActivity(shareIntent)
     }
 }
