+++ app/src/main/res/layout/activity_conversation_info.xml
@@ -229,7 +229,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:padding="@dimen/standard_padding"
-                    android:text="@string/message_expiration_title"
+                    android:text="@string/nc_conversation_settings"
                     android:textSize="@dimen/headline_text_size"
                     android:textStyle="bold" />
 
@@ -261,6 +261,36 @@
                     android:text="@string/nc_expire_messages_explanation"
                     android:textColor="@color/disabled_text"
                     android:textSize="@dimen/supporting_text_text_size" />
+
+            <LinearLayout
+                android:id="@+id/share_conversation_button"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:paddingStart="@dimen/standard_margin"
+                android:paddingTop="@dimen/standard_half_margin"
+                android:paddingEnd="@dimen/standard_margin"
+                android:paddingBottom="@dimen/standard_half_margin"
+                android:orientation="horizontal"
+                android:background="?android:attr/selectableItemBackground">
+
+
+                <ImageView
+                    android:layout_width="24dp"
+                    android:layout_height="40dp"
+                    android:layout_marginEnd="@dimen/standard_margin"
+                    android:contentDescription="@null"
+                    android:src="@drawable/ic_share_variant"
+                    app:tint="@color/grey_600" />
+
+                <com.google.android.material.textview.MaterialTextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="match_parent"
+                    android:gravity="center_vertical"
+                    android:text="@string/nc_guest_access_share_link"
+                    android:textSize="@dimen/headline_text_size" />
+
+            </LinearLayout>
+
             </LinearLayout>
 
             <LinearLayout
